Refactor features into modular files and add business logic tests
