+++ index.html
@@ -970,6 +970,18 @@
 <script src="js/core.js"></script>
 <script src="js/managers.js"></script>
 <script src="js/demo-data.js"></script>
+<script src="js/features/state.js"></script>
+<script src="js/features/utils.js"></script>
+<script src="js/features/filters.js"></script>
+<script src="js/features/cache.js"></script>
+<script src="js/features/mappings.js"></script>
+<script src="js/features/requests.js"></script>
+<script src="js/features/scenarios.js"></script>
+<script src="js/features/recording.js"></script>
+<script src="js/features/management.js"></script>
+<script src="js/features/request-api.js"></script>
+<script src="js/features/near-misses.js"></script>
+<script src="js/features/wiremock-extras.js"></script>
 <script src="js/features.js"></script>
 <script src="js/editor.js"></script>
 <script src="js/main.js"></script>
